+++ src/main/java/org/ojalgo/matrix/store/PhysicalStore.java
@@ -124,10 +124,6 @@ public interface PhysicalStore<N extends Comparable<N>> extends MatrixStore<N>, 
             return new RowsSupplier<>(this, numberOfColumns);
         }
 
-        default SparseColumnsStore<N> makeSparseColumns(final int nbRows, final int nbCols) {
-            return new SparseColumnsStore<>(this, nbRows, nbCols);
-        }
-
         default MatrixStore<N> makeSingle(final N element) {
             return new SingleStore<>(this, element);
         }